add test, that executes payment
- move payment related step definitions into own folder
- [CORE-395]
